+++ buffyboard/config.c
@@ -141,91 +141,19 @@ static void find_files(const char *path, char ***found, int *num_found) {
 }
 
 static int parsing_handler(void* user_data, const char* section, const char* key, const char* value) {
-    ul_config_opts *opts = (ul_config_opts *)user_data;
+    bb_config_opts *opts = (bb_config_opts *)user_data;
 
     if (strcmp(section, "general") == 0) {
-        if (strcmp(key, "animations") == 0) {
-            if (parse_bool(value, &(opts->general.animations))) {
-                return 1;
-            }
-        } else if (strcmp(key, "backend") == 0) {
-            ul_backends_backend_id_t id = ul_backends_find_backend_with_name(value);
-            if (id != UL_BACKENDS_BACKEND_NONE) {
-                opts->general.backend = id;
-                return 1;
-            }
-        } else if (strcmp(key, "timeout") == 0) {
-            /* Use a max ceiling of 60 minutes (3600 secs) */
-            opts->general.timeout = (uint16_t)LV_MIN(strtoul(value, (char **)NULL, 10), 3600);
-            return 1;
-        }
-    } else if (strcmp(section, "keyboard") == 0) {
-        if (strcmp(key, "autohide") == 0) {
-            if (parse_bool(value, &(opts->keyboard.autohide))) {
-                return 1;
-            }
-        } else if (strcmp(key, "layout") == 0) {
-            sq2lv_layout_id_t id = sq2lv_find_layout_with_short_name(value);
-            if (id != SQ2LV_LAYOUT_NONE) {
-                opts->keyboard.layout_id = id;
-                return 1;
-            }
-        } else if (strcmp(key, "popovers") == 0) {
-            if (parse_bool(value, &(opts->keyboard.popovers))) {
-                return 1;
-            }
-        }
-    } else if (strcmp(section, "textarea") == 0) {
-        if (strcmp(key, "obscured") == 0) {
-            if (parse_bool(value, &(opts->textarea.obscured))) {
-                return 1;
-            }
-        } else if (strcmp(key, "bullet") == 0) {
-            char *bullet = strdup(value);
-            if (bullet) {
-                opts->textarea.bullet = bullet;
-                return 1;
-            }
-        }
-    } else if (strcmp(section, "theme") == 0) {
-        if (strcmp(key, "default") == 0) {
+        if (strcmp(key, "theme") == 0) {
             bb_themes_theme_id_t id = bb_themes_find_theme_with_name(value);
             if (id != BB_THEMES_THEME_NONE) {
-                opts->theme.default_id = id;
-                return 1;
-            }
-        } else if (strcmp(key, "alternate") == 0) {
-            bb_themes_theme_id_t id = bb_themes_find_theme_with_name(value);
-            if (id != BB_THEMES_THEME_NONE) {
-                opts->theme.alternate_id = id;
-                return 1;
-            }
-        }
-    } else if (strcmp(section, "input") == 0) {
-        if (strcmp(key, "keyboard") == 0) {
-            if (parse_bool(value, &(opts->input.keyboard))) {
-                return 1;
-            }
-        } else if (strcmp(key, "pointer") == 0) {
-            if (parse_bool(value, &(opts->input.pointer))) {
-                return 1;
-            }
-        } else if (strcmp(key, "touchscreen") == 0) {
-            if (parse_bool(value, &(opts->input.touchscreen))) {
+                opts->general.theme_id = id;
                 return 1;
             }
         }
     } else if (strcmp(section, "quirks") == 0) {
         if (strcmp(key, "fbdev_force_refresh") == 0) {
             if (parse_bool(value, &(opts->quirks.fbdev_force_refresh))) {
-                return 1;
-            }
-        } else if (strcmp(key, "terminal_prevent_graphics_mode") == 0) {
-            if (parse_bool(value, &(opts->quirks.terminal_prevent_graphics_mode))) {
-                return 1;
-            }
-        } else if (strcmp(key, "terminal_allow_keyboard_input") == 0) {
-            if (parse_bool(value, &(opts->quirks.terminal_allow_keyboard_input))) {
                 return 1;
             }
         }
@@ -254,26 +182,12 @@ static bool parse_bool(const char *value, bool *result) {
  * Public functions
  */
 
-void ul_config_init_opts(ul_config_opts *opts) {
-    opts->general.animations = false;
-    opts->general.backend = ul_backends_backends[0] == NULL ? UL_BACKENDS_BACKEND_NONE : 0;
-    opts->general.timeout = 0;
-    opts->keyboard.autohide = true;
-    opts->keyboard.layout_id = SQ2LV_LAYOUT_US;
-    opts->keyboard.popovers = true;
-    opts->textarea.obscured = true;
-    opts->textarea.bullet = LV_SYMBOL_BULLET;
-    opts->theme.default_id = BB_THEMES_THEME_BREEZY_DARK;
-    opts->theme.alternate_id = BB_THEMES_THEME_BREEZY_LIGHT;
-    opts->input.keyboard = true;
-    opts->input.pointer = true;
-    opts->input.touchscreen = true;
+void bb_config_init_opts(bb_config_opts *opts) {
+    opts->general.theme_id = BB_THEMES_THEME_BREEZY_DARK;
     opts->quirks.fbdev_force_refresh = false;
-    opts->quirks.terminal_prevent_graphics_mode = false;
-    opts->quirks.terminal_allow_keyboard_input = false;
 }
 
-void ul_config_parse_directory(const char *path, ul_config_opts *opts) {
+void bb_config_parse_directory(const char *path, bb_config_opts *opts) {
     /* Find files in directory */
     char **found = NULL;
     int num_found = 0;
@@ -281,7 +195,7 @@ void ul_config_parse_directory(const char *path, ul_config_opts *opts) {
 
     /* Sort and parse files */
     qsort(found, num_found, sizeof(char *), compare_strings);
-    ul_config_parse_files((const char **)found, num_found, opts);
+    bb_config_parse_files((const char **)found, num_found, opts);
 
     /* Free memory */
     for (int i = 0; i < num_found; ++i) {
@@ -290,13 +204,13 @@ void ul_config_parse_directory(const char *path, ul_config_opts *opts) {
     free(found);
 }
 
-void ul_config_parse_files(const char **files, int num_files, ul_config_opts *opts) {
+void bb_config_parse_files(const char **files, int num_files, bb_config_opts *opts) {
     for (int i = 0; i < num_files; ++i) {
-        ul_config_parse_file(files[i], opts);
+        bb_config_parse_file(files[i], opts);
     }
 }
 
-void ul_config_parse_file(const char *path, ul_config_opts *opts) {
+void bb_config_parse_file(const char *path, bb_config_opts *opts) {
     bb_log(BB_LOG_LEVEL_VERBOSE, "Parsing config file %s", path);
     if (ini_parse(path, parsing_handler, opts) != 0) {
         bb_log(BB_LOG_LEVEL_ERROR, "Ignoring invalid config file %s", path);